Update draft 2020's and 2019's tests, so that they actually test features of each draft
