--- examples/both/test-2019SchemaDraft-invalid.json
@@ -1,14 +1,0 @@
-{
-    "schema": {
-        "$schema": "https://json-schema.org/draft/2019-09/schema",
-        "$id": "http://example.com/schema/test2019",
-        "type": "object",
-        "required": ["foo"],
-        "properties": {
-            "foo": { "type": "string" }
-        }
-    },
-    "data": {
-        "foo": 123
-    }
-}
