--- examples/both/test-2020SchemaDraft-invalid.json
@@ -1,14 +1,0 @@
-{
-    "schema": {
-        "$schema": "https://json-schema.org/draft/2020-12/schema",
-        "$id": "http://example.com/schema/test2020",
-        "type": "object",
-        "required": ["foo"],
-        "properties": {
-            "foo": { "type": "string" }
-        }
-    },
-    "data": {
-        "foo": 123
-    }
-}
