--- examples/objects/draft2019-9-support.json
@@ -1,10 +1,0 @@
-{
-  "sample_labels": [
-    {
-      "label": "test1"
-    },
-    {
-      "label": "test2"
-    }
-  ]
-}
