+++ examples/objects/draft2019-9-support_fail.json
@@ -1,0 +1,5 @@
+{
+  "a": "hello",
+  "b": 5,
+  "c": "invalid!"
+}
