+++ examples/objects/draft2019-9-support_pass.json
@@ -1,0 +1,4 @@
+{
+  "a": "hello",
+  "b": 5
+}
