+++ examples/objects/draft2020-12-support_fail.json
@@ -1,0 +1,1 @@
+["first", 42, "invalid!"]
