+++ examples/objects/draft2020-12-support_pass.json
@@ -1,0 +1,1 @@
+["first", 42, true, false]
