+++ examples/schemas/draft2019-9-support-schema.json
@@ -1,24 +1,25 @@
 {
   "$schema": "https://json-schema.org/draft/2019-09/schema",
-  "$id": "test_schema_X",
   "type": "object",
-  "required": ["sample_labels"],
-  "additionalProperties": false,
-  "properties": {
-    "sample_labels": {
-      "type": "array",
-      "minItems": 1,
-      "additionalProperties": false,
-      "items": {
-        "type": "object",
-        "required": ["label"],
-        "additionalProperties": false,
-        "properties": {
-          "label": {
-            "type": "string"
-          }
+  "unevaluatedProperties": false,
+  "$comment": "'a' is required, and 'b' is evaluated. If only 'a' or 'b' is present, the object is valid. If another property (e.g., 'c') not defined in these subschemas is present, the object is invalid.",
+  "allOf": [
+    {
+      "properties": {
+        "a": {
+          "type": "string"
+        }
+      },
+      "required": [
+        "a"
+      ]
+    },
+    {
+      "properties": {
+        "b": {
+          "type": "number"
         }
       }
     }
-  }
+  ]
 }
